sync

+++ modules/lev2-util/util/DateUtil.jl
@@ -39,8 +39,8 @@ export timult, calcRate
 bdaysPerYear() = 252.0
 durRisk(from::Date, to::Date) = 1 + bdays(from, to)
 timult(from::Date, to::Date) = bdaysPerYear() / durRisk(from, to)
-calcRate(from::Date, to::Date, ret, risk) = (ret / risk) * timult(from, to)
-calcRate(tmult, ret, risk) = (Float64(ret) / Float64(risk)) * tmult
+calcRate(from::Date, to::Date, ret, risk)::Float64 = (ret / Float64(risk)) * timult(from, to)
+calcRate(tmult, ret, risk)::Float64 = (ret / Float64(risk)) * tmult
 vixToSdev(vix, bdays) = vix / sqrt(bdaysPerYear() / bdays)
 
 #region Basic
